% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createReportFolder}
\alias{createReportFolder}
\title{Create a folder structure for a FEMO report.}
\usage{
createReportFolder(Folder = getwd(), ProjectName)
}
\arguments{
\item{Folder}{A folder to store the report structure. Defaults to your working directory.}

\item{ProjectName}{The name of the project. This will be used to name the folder in which the structure is created.}
}
\value{
Nested structures to ease data management.
}
\description{
\code{createReportFolder} creates the barebones of a FEMO report
}
\examples{
createReportFolder(Folder = "FEMO_Reports/", ProjectName = "BearMountainFire")
}

