% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rhPlot.R
\name{rhPlot}
\alias{rhPlot}
\title{Process RH data and plot it over time.}
\usage{
rhPlot(RH, DateTime, RH_LowBoundary = NULL, RH_HighBoundary = NULL)
}
\arguments{
\item{RH}{A vector of relative humidities}

\item{DateTime}{A vector of POSIXct objects corresponding
to the time the data were collected in format yyyy:mm:dd hh:mm.}

\item{RH_LowBoundary}{A scalar value of the low boundary parameters for the RH. Defaults to NULL.}

\item{RH_HighBoundary}{A scalar value of the high boundary parameters for the RH. Defaults to NULL.}
}
\value{
A ggplot of RH over time.
}
\description{
\code{rhPlot} processes a vector of cardinal wind directions and returns degrees
}
\examples{
myRH <- c(64,58,70,76)
Time <- lubridate::ymd_hm(c("2015/02/15 18:00",
                           "2015/02/15 18:30",
                           "2015/02/15 18:45",
                           "2015/02/15 19:00"))
df <- data.frame(myRH,Time)

High <- 70
Low <- 20

rhPlot(RH = df$myRH,
       DateTime = df$Time,
       RH_LowBoundary = Low,
       RH_HighBoundary = High)
}

