% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tempPlot.R
\name{tempPlot}
\alias{tempPlot}
\title{Process temperature data and plot it over time.}
\usage{
tempPlot(Temp, DateTime, Temp_LowBoundary = NULL, Temp_HighBoundary = NULL)
}
\arguments{
\item{Temp}{A vector of relative humidities}

\item{DateTime}{A vector of POSIXct objects
corresponding to the time the data were collected in format yyyy:mm:dd hh:mm.}

\item{Temp_LowBoundary}{A scalar value of the low boundary parameters for the RH. Defaults to NULL.}

\item{Temp_HighBoundary}{A scalar value of the high boundary parameters for the RH. Defaults to NULL.}
}
\value{
A ggplot of RH over time.
}
\description{
\code{tempPlot} processes a vector of cardinal wind directions and returns degrees
}
\examples{
myTemp <- c(72,68,52,48)
Time <- lubridate::ymd_hm(c("2015/02/15 18:00",
                           "2015/02/15 18:30",
                           "2015/02/15 18:45",
                           "2015/02/15 19:00"))
df <- data.frame(myTemp,Time)

High <- 70
Low <- 50

tempPlot(Temp = df$myTemp,
         DateTime = df$Time,
         Temp_LowBoundary = Low,
         Temp_HighBoundary = High)
}

