% Generated by roxygen2 (4.0.1): do not edit by hand
\name{windConvert}
\alias{windConvert}
\title{Process a vector of cardinal wind directions}
\usage{
windConvert(cardinalWind = vector(), LightAndVariable = "L/V")
}
\arguments{
\item{cardinalWind}{A vector of cardinal wind directions (i.e., NNE, SSE, SSW)}

\item{LightAndVariable}{A string that designates light and variable wind values. Default is "L/V"}
}
\value{
A vector of values given the degree angle of the wind.
}
\description{
\code{windConvert} processes a vector of cardinal wind directions and returns degrees
}
\examples{
windCardinal <- c("SE","SSE","NE","L/V")
windDegrees <- windConvert(windCardinal)
}

