% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/windPlot.R
\name{windPlot}
\alias{windPlot}
\title{Read wind data and return a plot of wind speed over time.}
\usage{
windPlot(DateTime, windDegrees, windSpeed, Wind_LowBoundary = NULL,
  Wind_HighBoundary = NULL)
}
\arguments{
\item{DateTime}{A vector of POSIXct objects corresponding
to the time the data were collected in format yyyy:mm:dd hh:mm.}

\item{windDegrees}{A vector of degree wind directions (i.e., 30, 45, 48)}

\item{windSpeed}{A string that designates light and variable wind values. Default is "L/V"}

\item{Wind_LowBoundary}{A scalar value of the low boundary parameters for the wind Defaults to NULL.}

\item{Wind_HighBoundary}{A scalar value of the high boundary parameters for the wind Defaults to NULL.}
}
\value{
A ggplot of wind speed and direction over time.
}
\description{
\code{windPlot} processes a vector of cardinal wind directions and returns degrees
}
\examples{
windCardinal <- c("SE","SSE","NE","L/V")
Degrees <- windConvert(windCardinal, LightAndVariable = "L/V")
Speed <- c(5, 2 , 3, NA) # Replace light and variable speeds with NA
Time <- lubridate::ymd_hm(c("2015/02/15 18:00",
                           "2015/02/15 18:30",
                           "2015/02/15 18:45",
                           "2015/02/15 19:00"))
windPlot(Time,
         Degrees,
         Speed,
         Wind_LowBoundary = 1,
         Wind_HighBoundary = 10)
}

