% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createReportFolder.R
\name{createReportFolder}
\alias{createReportFolder}
\title{Create a folder structure for a FEMO report.}
\usage{
createReportFolder(Folder = getwd(), ProjectName, ProjectDate)
}
\arguments{
\item{Folder}{A folder to store the report structure. Defaults to your working directory.}

\item{ProjectName}{The name of the project.
This will be used to name the folder in which the structure is created.}

\item{ProjectDate}{The data of the project in 'YYYYMMDD'}
}
\value{
Nested structures to ease data management.
}
\description{
\code{createReportFolder} creates the barebones of a FEMO report
}
\examples{
createReportFolder(Folder = "FEMO_Reports/",
                   ProjectName = "BearMountainFire",
                   ProjectDate = 20150716)
}

