% Generated by roxygen2 (4.0.1): do not edit by hand
\name{windPlot}
\alias{windPlot}
\title{Process wind data and return a plot of wind speed over time.}
\usage{
windPlot(DateTime, windDegrees, windSpeed)
}
\arguments{
\item{windDegrees}{A vector of degree wind directions (i.e., 30, 45, 48)}

\item{DateTime}{A vector of POSIXct objects corresponding to the time the data were collected in format yyyy:mm:dd hh:mm.}

\item{windSpeed}{A string that designates light and variable wind values. Default is "L/V"}
}
\value{
A ggplot of wind speed and direction over time.
}
\description{
\code{windPlot} processes a vector of cardinal wind directions and returns degrees
}
\examples{
windCardinal <- c("SE","SSE","NE","L/V")
Degrees <- windConvert(windCardinal, LightAndVariable = "L/V")
Speed <- c(5, 2 , 3, "L/V")
Time <- ymd_hm(c("2015/02/15 18:00","2015/02/15 18:30","2015/02/15 18:45","2015/02/15 19:00"))
}

